f17: move settings.py to Model/Settings.py
